pyproject.toml
[tool.poetry]
name = "llm-moderator"
version = "0.1.0"
description = "LLM-based content moderation project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "llm_moderator"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
torch = "^2.9.1"
pytorch-lightning = "^2.6.0"
hydra-core = "^1.3.2"
pandas = "^2.3.3"
scikit-learn = "^1.7.2"
pyyaml = "^6.0.3"
transformers = "^4.57.3"
fire = "^0.6.0"
mlflow = "^2.16.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.5.0"
ruff = "^0.14.8"
black = "^25.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
src = ["llm_moderator"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.poetry.scripts]
llm-cli = "llm_moderator.cli:run"
